pyproject.toml
[project]
name = "harness-engineering"
version = "0.1.0"
description = "GitHub-control-plane implementation of the Symphony orchestration service."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.2",
]

[project.scripts]
symphony-harness = "harness_engineering.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
  "mypy>=1.11",
  "pre-commit>=3.8",
  "pytest>=8.0",
  "ruff>=0.6.9",
  "types-PyYAML>=6.0.12",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[tool.ruff]
target-version = "py311"
line-length = 140
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = [
  "E501",
]

[tool.ruff.lint.isort]
known-first-party = ["harness_engineering"]

[tool.mypy]
python_version = "3.11"
files = ["src/harness_engineering"]
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unreachable = true
pretty = true
